+++ tests/tests/models/test_collection.py
@@ -346,7 +346,7 @@ class CollectionItemMixin:
             self.resource_extension  # noqa: F821
         )
 
-    def test_file_supported(self, storage):
+    def test_accept(self, storage):
         raise NotImplementedError
 
 
@@ -423,18 +423,18 @@ class TestFileItem(CollectionItemMixin, TestFileFieldResource):
             'uploaded': storage.resource.uploaded_at.isoformat(),
         }
 
-    def test_file_supported(self, storage):
+    def test_accept(self, storage):
         with open(DOCUMENT_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is True
+            assert storage.resource.accept(File(fp)) is True
 
         with open(NATURE_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is True
+            assert storage.resource.accept(File(fp)) is True
 
         with open(MEDITATION_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is True
+            assert storage.resource.accept(File(fp)) is True
 
         with open(AUDIO_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is True
+            assert storage.resource.accept(File(fp)) is True
 
 
 @pytest.mark.django_db
@@ -625,18 +625,18 @@ class TestSVGItem(CollectionItemMixin, TestFileFieldResource):
         with storage.resource.open() as fp:
             assert fp.read(5) == b'<?xml'
 
-    def test_file_supported(self, storage):
+    def test_accept(self, storage):
         with open(DOCUMENT_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is False
+            assert storage.resource.accept(File(fp)) is False
 
         with open(NATURE_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is False
+            assert storage.resource.accept(File(fp)) is False
 
         with open(MEDITATION_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is True
+            assert storage.resource.accept(File(fp)) is True
 
         with open(AUDIO_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is False
+            assert storage.resource.accept(File(fp)) is False
 
 
 @pytest.mark.django_db
@@ -852,19 +852,19 @@ class TestImageItem(CollectionItemMixin, TestFileFieldResource):
     def test_height(self, storage):
         assert storage.resource.height == 2301
 
-    def test_file_supported(self, storage):
+    def test_accept(self, storage):
         with open(DOCUMENT_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is False
+            assert storage.resource.accept(File(fp)) is False
 
         with open(NATURE_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is True
+            assert storage.resource.accept(File(fp)) is True
 
         # SVG passes image test
         with open(MEDITATION_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is True
+            assert storage.resource.accept(File(fp)) is True
 
         with open(AUDIO_FILEPATH, 'rb') as fp:
-            assert storage.resource.file_supported(File(fp)) is False
+            assert storage.resource.accept(File(fp)) is False
 
 
 @pytest.mark.django_db